```text Refactor code to cast if_lims as float for improved data handling
The `if_lims` variable in the `NYGrid` class has been modified to cast the values as float for improved data handling. This change ensures that the calculations involving `if_lims` are performed accurately.

+++ nygrid/nygrid.py
@@ -516,7 +516,7 @@ class NYGrid:
 
         # Get interface limit information
         self.if_map = self.ppc_int['if']['map']
-        self.if_lims = self.ppc_int['if']['lims']
+        self.if_lims = self.ppc_int['if']['lims'].astype(float)
         self.if_lims[:, 1:] = self.if_lims[:, 1:] / self.baseMVA
         self.NIF = len(self.if_lims)
